requirements.txt
django 
psycopg2
tensorflow
torch
sentence_transformers
pandas
tensorflow_hub
djangorestframework